Created sessions Reward creation integration test. updated Bulk insertion insertion. introduced preload extention to load the user in one standardized way.

+++ src/RewardFlow.API/Rewards/SessionsReward/Dtos/SessionRewardDto.cs
@@ -4,7 +4,7 @@ public record SessionRewardDto(
     int SessionRewardId,
     string Name,
     string? Code,
-    byte? Year,
+    short? Year,
     byte? Semester,
     decimal Percentage,
     decimal Total,
